+++ music/src/main/java/com/music/upload/convert/AudioQualityConverter.java
@@ -1,4 +1,4 @@
-package com.music.upload.service;
+package com.music.upload.convert;
 
 import com.music.adaptor.AudioConverter;
 import com.music.constatns.AudioQuality;
